% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning-params.R
\name{calcBalance}
\alias{calcBalance}
\title{Weighted covariate balance}
\usage{
calcBalance(x.mat, trt1, trt2, weights, d.type = 2)
}
\arguments{
\item{x.mat}{covariate matrix.}

\item{trt1}{indices of treated units.}

\item{trt2}{indices of control units.}

\item{weights}{propensity score (IPTW) weights.}

\item{d.type}{standard deviation used: 1 = treated group, 2 = pooled.}
}
\value{
Matrix of absolute standardised mean differences, one per covariate.
}
\description{
Absolute standardised differences in weighted covariate means between
treated and control units.
}
