% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spbalance-package.R
\docType{package}
\name{spbalance-package}
\alias{spbalance}
\alias{spbalance-package}
\title{spbalance: covariate balancing propensity scores with spatial confounding}
\description{
Estimates propensity scores with a loss that prioritises covariate balance
(\code{\link[=spBalance]{spBalance()}}), allowing spline and kernel terms in space to adjust for
unmeasured spatial confounding, and estimates the average treatment effect
with iid, spatial HAC or spatial block bootstrap standard errors
(\code{\link[=spbalATE]{spbalATE()}}).
}
\author{
\strong{Maintainer}: Nathan B. Wikle \email{nbwikle@gmail.com}

}
\keyword{internal}
